requirements.txt
fastapi~=0.68.0
uvicorn~=0.15.0
pydantic~=1.8.0
python-multipart~=0.0.5
aiofiles~=0.8.0
PyYAML~=6.0.0
docker~=6.0.0
aiohttp~=3.8.0
croniter~=1.3.0
pytz~=2023.3
websockets~=10.0.0
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
SQLAlchemy~=1.4.0
alembic~=1.7.0
aiosqlite~=0.17.0
python-dotenv~=0.19.0
watchdog~=2.1.0
asyncpg~=0.27.0
aiomysql~=0.1.1
pymysql~=1.0.2
cryptography~=3.4.7
greenlet~=1.1.0 